refactored paths, hopefully functional model train/save functionality, added some more documentation, renamed things for easier to read github
