+++ Pipeline/s0_session_manager.py
